requirements.txt
django==3.2.9
pillow==8.4.0
pyJWT==2.3.0
coverage==6.1.1
django-debug-toolbar==3.2.2
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
requests==2.26.0
drf-yasg==1.20.0
psycopg2==2.9.5
social-auth-app-django==4.0.0